# openapi.yaml
openapi: 3.1.0
info:
  title: PromptSpeak Symbol Registry API
  description: |
    REST API for managing PromptSpeak symbols. Symbols are structured context frames
    that ground LLM responses with verified facts, requirements, and constraints.

    ## Authentication
    All endpoints (except /health) require an API key passed in the `X-API-Key` header.

    ## Symbol ID Format
    Symbol IDs follow the pattern: `Ξ.<TICKER>.<CONTEXT>` (e.g., `Ξ.NVDA.Q3FY25`)

    ## Use with CustomGPT
    1. Add this spec as an Action in your CustomGPT
    2. Configure authentication with your API key
    3. Use symbols to ground responses with verified context
  version: 1.0.0
  contact:
    name: PromptSpeak API Support

servers:
  - url: https://your-domain.com
    description: Production server (replace with your deployed URL)
  - url: http://localhost:3000
    description: Local development server

security:
  - ApiKeyAuth: []

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key for authentication (format: ps_live_xxxxxxxxxxxx)"

  schemas:
    Symbol:
      type: object
      required:
        - symbolId
        - who
        - what
        - why
        - where
        - when
        - commanders_intent
        - requirements
      properties:
        symbolId:
          type: string
          description: Unique identifier starting with "Ξ."
          example: "Ξ.NVDA.Q3FY25"
        version:
          type: integer
          description: Symbol version number
          example: 1
        hash:
          type: string
          description: Content hash for integrity verification
          example: "4820bd39239ff28a"
        category:
          type: string
          enum: [COMPANY, PERSON, EVENT, SECTOR, TASK, KNOWLEDGE, QUERY]
          description: Symbol category (auto-detected if not specified)
        who:
          type: string
          description: The subject or entity
          example: "NVIDIA Corporation"
        what:
          type: string
          description: What is being analyzed or tracked
          example: "Q3 FY25 Earnings Analysis"
        why:
          type: string
          description: Purpose or motivation
          example: "Track AI chip market leadership"
        where:
          type: string
          description: Context or domain
          example: "US semiconductor sector"
        when:
          type: string
          description: Time reference or period
          example: "2024-11-20"
        how:
          type: object
          properties:
            focus:
              type: array
              items:
                type: string
              description: Areas to focus on
              example: ["datacenter", "gaming", "automotive"]
            constraints:
              type: array
              items:
                type: string
              description: Constraints for analysis
              example: ["cite official sources", "include YoY comparisons"]
            output_format:
              type: string
              description: Preferred output format
        commanders_intent:
          type: string
          description: High-level goal or intent
          example: "Monitor NVIDIA financial performance for AI investment thesis"
        requirements:
          type: array
          items:
            type: string
          description: Required elements in responses
          example: ["Revenue breakdown by segment", "Data center growth metrics"]
        anti_requirements:
          type: array
          items:
            type: string
          description: Elements to exclude from responses
        key_terms:
          type: array
          items:
            type: string
          description: Important terminology
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        created_by:
          type: string
          description: Creator user ID
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        changelog:
          type: array
          items:
            type: object
            properties:
              version:
                type: integer
              change:
                type: string
              timestamp:
                type: string
                format: date-time
              changed_by:
                type: string

    CreateSymbolRequest:
      type: object
      required:
        - symbolId
        - who
        - what
        - why
        - where
        - when
        - commanders_intent
        - requirements
      properties:
        symbolId:
          type: string
          description: Unique identifier starting with "Ξ."
          example: "Ξ.NVDA.Q3FY25"
        who:
          type: string
          example: "NVIDIA Corporation"
        what:
          type: string
          example: "Q3 FY25 Earnings Analysis"
        why:
          type: string
          example: "Track AI chip market leadership"
        where:
          type: string
          example: "US semiconductor sector"
        when:
          type: string
          example: "2024-11-20"
        how:
          type: object
          properties:
            focus:
              type: array
              items:
                type: string
            constraints:
              type: array
              items:
                type: string
            output_format:
              type: string
        commanders_intent:
          type: string
          example: "Monitor NVIDIA financial performance for AI investment thesis"
        requirements:
          type: array
          items:
            type: string
          example: ["Revenue breakdown by segment", "Data center growth metrics"]
        anti_requirements:
          type: array
          items:
            type: string
        key_terms:
          type: array
          items:
            type: string

    UpdateSymbolRequest:
      type: object
      required:
        - changes
        - change_description
      properties:
        changes:
          type: object
          description: Fields to update (partial update supported)
          additionalProperties: true
        change_description:
          type: string
          description: Description of the change for changelog
          example: "Updated revenue projections"

    ImportSymbolsRequest:
      type: object
      required:
        - source
        - data
        - category
        - id_prefix
      properties:
        source:
          type: string
          description: Source system name
          example: "sec_filings"
        data:
          type: array
          items:
            type: object
          description: Array of symbol data to import
        category:
          type: string
          enum: [COMPANY, PERSON, EVENT, SECTOR, TASK, KNOWLEDGE, QUERY]
        id_prefix:
          type: string
          description: Prefix for generated symbol IDs
          example: "Ξ.SEC"
        transform:
          type: object
          description: Field mapping rules
        defaults:
          type: object
          description: Default values for missing fields

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Error code
        requestId:
          type: string
          description: Request ID for debugging

paths:
  /health:
    get:
      summary: Health check
      description: Check API health status (no authentication required)
      operationId: getHealth
      security: []
      tags:
        - System
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [healthy, degraded, unhealthy]
                  timestamp:
                    type: string
                    format: date-time
                  version:
                    type: string
                  uptime:
                    type: number

  /api/v1/symbols:
    get:
      summary: List symbols
      description: |
        Retrieve a list of symbols with optional filtering.
        Returns summary information for each symbol.
      operationId: listSymbols
      tags:
        - Symbols
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [COMPANY, PERSON, EVENT, SECTOR, TASK, KNOWLEDGE, QUERY]
          description: Filter by category
        - name: tags
          in: query
          schema:
            type: string
          description: Comma-separated list of tags to filter by
        - name: search
          in: query
          schema:
            type: string
          description: Search term for symbol content
        - name: created_after
          in: query
          schema:
            type: string
            format: date-time
          description: Filter symbols created after this date
        - name: created_before
          in: query
          schema:
            type: string
            format: date-time
          description: Filter symbols created before this date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Maximum number of results
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Offset for pagination
      responses:
        '200':
          description: List of symbols
          content:
            application/json:
              schema:
                type: object
                properties:
                  symbols:
                    type: array
                    items:
                      type: object
                      properties:
                        symbolId:
                          type: string
                        category:
                          type: string
                        version:
                          type: integer
                        hash:
                          type: string
                        commanders_intent:
                          type: string
                        created_at:
                          type: string
                          format: date-time
                  total:
                    type: integer
                  has_more:
                    type: boolean
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a symbol
      description: |
        Create a new symbol with the 5W+H framework.
        The symbol ID must start with "Ξ." and be unique.
      operationId: createSymbol
      tags:
        - Symbols
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSymbolRequest'
      responses:
        '201':
          description: Symbol created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  symbolId:
                    type: string
                  version:
                    type: integer
                  hash:
                    type: string
                  path:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Symbol already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/symbols/{symbolId}:
    get:
      summary: Get a symbol
      description: |
        Retrieve a symbol by ID. Optionally include changelog history.
        Symbol IDs contain dots, so they must be URL-encoded (Ξ = %CE%9E).
      operationId: getSymbol
      tags:
        - Symbols
      parameters:
        - name: symbolId
          in: path
          required: true
          schema:
            type: string
          description: Symbol ID (URL-encoded, e.g., %CE%9E.NVDA.Q3FY25)
          example: "Ξ.NVDA.Q3FY25"
        - name: version
          in: query
          schema:
            type: integer
          description: Specific version to retrieve (default is latest)
        - name: include_changelog
          in: query
          schema:
            type: boolean
            default: false
          description: Include version history
      responses:
        '200':
          description: Symbol found
          content:
            application/json:
              schema:
                type: object
                properties:
                  found:
                    type: boolean
                  symbol:
                    $ref: '#/components/schemas/Symbol'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Symbol not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update a symbol
      description: |
        Update specific fields of a symbol. Creates a new version.
        Include a change description for the changelog.
      operationId: updateSymbol
      tags:
        - Symbols
      parameters:
        - name: symbolId
          in: path
          required: true
          schema:
            type: string
          description: Symbol ID (URL-encoded)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSymbolRequest'
      responses:
        '200':
          description: Symbol updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  symbolId:
                    type: string
                  old_version:
                    type: integer
                  new_version:
                    type: integer
                  old_hash:
                    type: string
                  new_hash:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Symbol not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a symbol
      description: |
        Delete a symbol from the registry.
        Optionally provide a reason for the deletion.
      operationId: deleteSymbol
      tags:
        - Symbols
      parameters:
        - name: symbolId
          in: path
          required: true
          schema:
            type: string
          description: Symbol ID (URL-encoded)
        - name: reason
          in: query
          schema:
            type: string
          description: Reason for deletion
      responses:
        '200':
          description: Symbol deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deleted:
                    type: boolean
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Symbol not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/symbols/stats:
    get:
      summary: Get registry statistics
      description: Get aggregate statistics about the symbol registry
      operationId: getStats
      tags:
        - Symbols
      responses:
        '200':
          description: Registry statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  registry_stats:
                    type: object
                    properties:
                      total_symbols:
                        type: integer
                      by_category:
                        type: object
                        additionalProperties:
                          type: integer
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/symbols/import:
    post:
      summary: Bulk import symbols
      description: |
        Import multiple symbols from an external source.
        Supports field transformation and default values.
      operationId: importSymbols
      tags:
        - Symbols
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportSymbolsRequest'
      responses:
        '201':
          description: Import completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  imported:
                    type: integer
                  failed:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        index:
                          type: integer
                        error:
                          type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

tags:
  - name: System
    description: System health and status endpoints
  - name: Symbols
    description: Symbol CRUD operations